+++ Tests/Elements/Button.cs
@@ -6,11 +6,11 @@ namespace Tests.Elements
 
     public class Button: Element, IClickable
     {
-        public Button(string locator, [CallerMemberName]string name = null) : base(locator, name)
+        public Button(string locator, [CallerMemberName]string name = "") : base(locator, name)
         {
         }
 
-        public Button(Element source, [CallerMemberName]string name = null, params object[] locatorFormattingArguments) : base(source, name, locatorFormattingArguments)
+        public Button(Element source, [CallerMemberName]string name = "", params object[] locatorFormattingArguments) : base(source, name, locatorFormattingArguments)
         {
         }
     }