+++ Tests/Elements/Textbox.cs
@@ -6,11 +6,11 @@ namespace Tests.Elements
 
     public class Textbox : Element, ITypeable
     {
-        public Textbox(string locator, [CallerMemberName]string name = null) : base(locator, name)
+        public Textbox(string locator, [CallerMemberName]string name = "") : base(locator, name)
         {
         }
-        
-        public Textbox(Element source, [CallerMemberName]string name = null, params object[] locatorFormattingArguments) : base(source, name, locatorFormattingArguments)
+
+        public Textbox(Element source, [CallerMemberName]string name = "", params object[] locatorFormattingArguments) : base(source, name, locatorFormattingArguments)
         {
         }
     }